common/spi: Rewrite SpiBuilder
since the previous use of generics wouldn't allow creating a Builder
instance with one of the SPI lines left out, for example. Instead, do
not collect the IO lines in the Builder struct, but directly `connect`
them with their respective peripherals when the respective functions are
called.

This essentially "allocates" the IO even before the finished SPI
instance is collected from the builder. That is possible in this
instance because once the GPIO has been allocated via the `mosi/miso/cs`
functions, it isn't accessible to the user anymore due to the borrow
checker identifying it has been moved.