feat(esp32c3): gpio_interrupt example ISR in SRAM
The culmination of this entire patch series, we now place the interrupt
handling code itself into SRAM. The `#[ram]` attribute is for the `GPIO`
function itself (and `#[inline(never)]` works fine here because LLVM
can't "see" through the ~indirect reference to this function), but then
that leaves the matter of the function body itself.

Unfortunately, this is necessarily work specific to the details of each
handler's implementation. Elsewhere, I got started down this road
because we needed SPI to start up within a certain time bound, and so
we needed to add attributes to much of the HAL's SPI implementation.
And, in addition to `memset` and `memcpy` we had to override `memmove`
as well, because something about the datastructures used for DMA made
`rustc` use that in a few places.

As far as this particular example goes, touching `core::fmt` pulls in a
lot of code from fairly far outside our known methods of control:

```
        write(&mut self, args)
40380f00:       3c006537                lui     a0,0x3c006
40380f04:       7a850513                addi    a0,a0,1960 # 3c0067a8 <.Lanon.6ccc2de621f5491d2d69154e15b03ba0.16>
40380f08:       fca42c23                sw      a0,-40(s0)
40380f0c:       4505                    li      a0,1
40380f0e:       fca42e23                sw      a0,-36(s0)
40380f12:       3c006537                lui     a0,0x3c006
40380f16:       6a450513                addi    a0,a0,1700 # 3c0066a4 <anon.f0eba740dd2cb5e3ab050cbaac366ec5.84.llvm
.5894916478550538780>
40380f1a:       fea42023                sw      a0,-32(s0)
40380f1e:       fe042223                sw      zero,-28(s0)
40380f22:       fe042423                sw      zero,-24(s0)
40380f26:       3c006537                lui     a0,0x3c006
40380f2a:       5fc50593                addi    a1,a0,1532 # 3c0065fc <anon.7f39784d2cd27331f7ef17a88d595274.0.llvm.
10122694886931673864>
40380f2e:       fd440513                addi    a0,s0,-44
40380f32:       fd840613                addi    a2,s0,-40
40380f36:       01c82097                auipc   ra,0x1c82
40380f3a:       92a080e7                jalr    -1750(ra) # 42002860 <core::fmt::write>
```

So, this change circumvents the convenience machinery to call
`write_str` directly, thereby avoiding touching `core::fmt::write` and
its anonymous helper data. Curiously, with the combination of
optimization settings in use, doing so actually erases the `"GPIO
Interrupt\n"` string entirely in favor of an inlined, unrolled loop
emitting each byte in succession:

```asm
40380efe:       40000537                lui     a0,0x40000
40380f02:       06850493                addi    s1,a0,104 # 40000068 <ets_delay_us+0x18>
40380f06:       04700513                li      a0,71
40380f0a:       9482                    jalr    s1
40380f0c:       05000513                li      a0,80
40380f10:       9482                    jalr    s1
...
```

(for those, like me, who have not memorized the decimal->ACII lookup
table, 71 is 'G' and 80 is 'P')

NB: LTO is _required_ at this stage: turning it off produces calls to
things like `_critical_section_1_0_{acquire,release}` and some `Mutex`
interactions. The latter might be fixable, but I didn't look into it
because the former is seemingly not: the `extern "Rust"` seems to form a
boundary that currently can only be pierced by applying LTO.