RMT: Refactor structs/traits/channel creation to avoid macros
This commit completely rearranges the internal driver structure, whereas
the user-facing interface should be mostly unchanged (some visible trait
bounds might differ).

Crucially, this adds the `RawChannelAccess` trait, which is implemented
by `ConstChannelAccess` and the type-erased variant `DynChannelAccess`:
The latter provide low-level access to a given channel (specified either
via const generics, or dynamically). Ownership of these types or the
`ChannelCreator`s indicate exclusive access to the given channel (with
the caveat of extended channel RAM, which might reserve some channels,
which is checked at runtime upon channel configuration, as before this
commit).

Marker structs `Unconfigured`, `Tx`, `Rx` are used to indicate the
current configuration of a channel. Marker structs `Rx`, `Tx`
and `RxTx` indicate which capabilities a given channel has, and thereby
which configuration a given `ChannelCreator` allows for.

Overall, we have:

Marker structs
- Tx: Channel direction/capability
- Rx: Channel direction/capability
- RxTx: Channel capability

Low-level channel access tokens
- trait RawChannelAccess
- struct ConstChannelAccess (a ZST)
	- impl RawChannelAccess for ConstChannelAccess
- struct DynChannelAccess (u8-sized)
	- impl RawChannelAccess for DynChannelAccess

Low-level channel access methods
- trait ChannelInternal (This is new, bundling some methods that are
  valid for both Rx and Tx channels, and were previously duplicated in
  the other two traits)
- trait TxChannelInternal (mostly unchanged)
- trait RxChannelInternal (mostly unchanged)

which have blanket implementations for RawChannelAccess

User-facing types are `Rmt`, `Channel`, `ChannelCreator` as before.
Internally, `Channel`s do not implement the *ChannelInternal traits
directly anymore, but own a `RawChannelAccess` impl which provides them.
Thus, if the latter is a `ConstChannelAccess`, the compiler should be
able to const-propagate and inline all code as before. However, we can
now also have `Channel<Dm, DynChannelAccess<Tx|Rx>>` which don't
monomorphize and the channel number, but handle it dynamically.

This removes most of the macro usage, avoids some code duplication, and
enables type-erased channels. Chip-specific code is now confined to the
`declare_channels!` and `declare_signals!` macros (which declare the
precise number and capabilities of a chip's channels) and the
`chip_specific` modules (which implement low-level access to the two
variants of RMT peripherals that exist in current chips), centralizing
this more than before.